luxonis_ml/utils/requirements.txt
pydantic>=2.4.0
pydantic-settings>=2.1.0
PyYAML>=6.0
fsspec>=2023.1.0
rich>=13.6.0
# s3fs # if using S3 filesystem
# gcsfs # if using GCS filesystem
# mlflow # if using MLflow
